Return int64 for getSize to support > 2GB

+++ hi_scripting/scripting/api/ScriptingApiObjects.cpp
@@ -274,7 +274,7 @@ var ScriptingObjects::ScriptFile::createDirectory(String directoryName)
 	return new ScriptFile(getScriptProcessor(), f.getChildFile(directoryName));
 }
 
-int ScriptingObjects::ScriptFile::getSize()
+int64 ScriptingObjects::ScriptFile::getSize()
 {	
 	return f.getSize();
 }
